A New hero page with image created, old hero is renamed as workouts
