+++ client/src/App.js
@@ -1,35 +1,33 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { useAuthContext } from "./hooks";
-import { Home, LogIn, SignUp } from "./pages";
-import { Navbar } from "./components";
-import { Footer } from "./components";
+import { Home, Workouts, LogIn, SignUp } from "./pages";
+import { Navbar, Footer } from "./components";
 
 function App() {
   const { user } = useAuthContext();
 
   return (
-    <div className="App">
-      <BrowserRouter>
+    <BrowserRouter>
+      <div className="app">
         <Navbar />
-        <div className="pages">
-          <Routes>
-            <Route
-              path="/"
-              element={user ? <Home /> : <Navigate to="/login" />}
-            ></Route>
-            <Route
-              path="/login"
-              element={!user ? <LogIn /> : <Navigate to="/" />}
-            ></Route>
-            <Route
-              path="/signup"
-              element={!user ? <SignUp /> : <Navigate to="/" />}
-            ></Route>
-          </Routes>
-        </div>
+        <Routes>
+          <Route path="/" element={<Home />}></Route>
+          <Route
+            path="/workouts"
+            element={user ? <Workouts /> : <Navigate to="/login" />}
+          ></Route>
+          <Route
+            path="/login"
+            element={!user ? <LogIn /> : <Navigate to="/workouts" />}
+          ></Route>
+          <Route
+            path="/signup"
+            element={!user ? <SignUp /> : <Navigate to="/workouts" />}
+          ></Route>
+        </Routes>
         <Footer />
-      </BrowserRouter>
-    </div>
+      </div>
+    </BrowserRouter>
   );
 }
