+++ client/src/pages/index.js
@@ -1,5 +1,6 @@
 import Home from "./Home.js";
+import Workouts from "./Workouts.js";
 import LogIn from "./LogIn.js";
 import SignUp from "./SignUp.js";
 
-export { Home, LogIn, SignUp };
+export { Home, Workouts, LogIn, SignUp };
